requirements.txt
anthropic>=0.40.0
mcp>=1.0.0
pyyaml>=6.0
jsonschema>=4.0
typer>=0.12.0
httpx>=0.27.0
rich>=13.0.0
